Split channel_operations.py by command.
It was messy.
